// PlanItKorea_w/src/types/type.ts
export type Accommodation = "호텔&리조트" | "펜션&풀빌라" | "캠핑&글램핑";

export type Facilities =
  | "사우나"
  | "수영장"
  | "바베큐"
  | "세탁 가능"
  | "스파/월풀"
  | "와이파이"
  | "에어컨"
  | "욕실용품"
  | "샤워실"
  | "조식포함"
  | "무료주차"
  | "반려견 동반"
  | "객실 내 취사"
  | "OTT";